+++ src/MicroOcpp/Model/Model.cpp
@@ -17,6 +17,7 @@
 #include <MicroOcpp/Model/Reservation/ReservationService.h>
 #include <MicroOcpp/Model/Boot/BootService.h>
 #include <MicroOcpp/Model/Reset/ResetService.h>
+#include <MicroOcpp/Model/Certificates/CertificateService.h>
 
 #include <MicroOcpp/Core/Configuration.h>
 
@@ -185,6 +186,15 @@ void Model::setResetService(std::unique_ptr<ResetService> rs) {
 
 ResetService *Model::getResetService() const {
     return resetService.get();
+}
+
+void Model::setCertificateService(std::unique_ptr<CertificateService> cs) {
+    this->certService = std::move(cs);
+    capabilitiesUpdated = true;
+}
+
+CertificateService *Model::getCertificateService() const {
+    return certService.get();
 }
 
 Clock& Model::getClock() {